% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email_notification.R
\name{send_email_notification}
\alias{send_email_notification}
\title{Send email notification}
\usage{
send_email_notification(to, subject, body, smtp_settings = smtp_settings())
}
\arguments{
\item{to}{to}

\item{subject}{subject}

\item{body}{body}

\item{smtp_settings}{smtp_settings as retrieved by \code{\link{smtp_settings}}}
}
\value{
sends email
}
\description{
Send email notification
}
