% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email_notification.R
\name{smtp_settings}
\alias{smtp_settings}
\title{SMTP configuration for email notification}
\usage{
smtp_settings()
}
\value{
list with smtp configuration
}
\description{
SMTP configuration for email notification
}
