% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_starting_places.R
\name{check_starting_places}
\alias{check_starting_places}
\title{ZUT25 check for available starting places for each track}
\usage{
check_starting_places(url, send_pushover_notification = FALSE, dbg = TRUE)
}
\arguments{
\item{url}{one element of \code{\link{zut25_urls}}}

\item{send_pushover_notification}{should pushover notifiation be sent (default: FALSE)}

\item{dbg}{print debug messages (default: TRUE)}
}
\value{
either 0 (if no starting place is available) or number of available
places as text. in additon an email will be sent to recipient email in case
a starting place is available
}
\description{
ZUT25 check for available starting places for each track
}
\examples{
sapply(zut25_urls(), check_starting_places)

}
